requirements.txt
Django
django-channels
django-cors-headers
django-extensions
django-sslserver
asgiref
async-timeout
certifi
cffi
cryptography
PyJWT
channels
channels-redis
charset-normalizer
click
djangorestframework
gunicorn
h11
MarkupSafe
msgpack
packaging
psycopg2-binary
pycparser
python-dotenv
pytz
redis
requests-oauthlib
sqlparse
typing_extensions
uvicorn
websockets
Werkzeug
